Make Obscure generic and Secret comparable
This commit uses generics with Obscure and Secret to make them
accept both string and byte slice types. It also makes Secret
comparable so it may be used as a map key.

The methods Equal and Compare have been dropped. The Equal method is
replaced by the == operator. The Compare method is obsoleted by most
users not needing to sort obscured secrets.